Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.4
tensorflow==2.13.0
opencv-python==4.8.1.78
ultralytics==8.0.206
requests==2.31.0
python-dotenv==1.0.0
reportlab==4.0.4
